+++ src/main/resources/templates/profile.html
@@ -12,7 +12,7 @@
 <img th:if="${user.avatar != null}" th:src="@{/uploads/avatar/{id}(id=${user.avatar.id})}" class="img-thumbnail"
      style="height:200px"/>
 
-<form th:action="@{/profile}" method="post" enctype="multipart/form-data"
+<form th:action="@{/profileAvatar}" method="post" enctype="multipart/form-data"
       class="mt-3 mb-5 pb-5 row-col-xs-3">
     <div class="form-group">
         <span th:if="${error}" class="alert alert-primary" role="alert">Неверный формат файла!</span>
